requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
httpx>=0.27.0
python-dotenv>=1.0.0
solders>=0.21.0
solana>=0.34.0
base58>=2.1.0
websockets>=12.0
pydantic>=2.0.0
